+++ NEWS.md
@@ -1,6 +1,6 @@
 # optparse NEWS - User visible changes
 
-## Noteworthy changes in release ?.? (????-??-??) [?]
+## Noteworthy changes in release 1.1 (2016-01-29) [stable]
 
 ### New features